fix: `kube-aws up --s3-uri s3://<bucket>/<directory> to automatically avoid the 51200 bytes limitation errors of CloudFormation
If you are hit by the cloudformation limit, `kube-aws up` now fail with a specific error explaining the limit and how to work-around it:

```
$ kube-aws up
Creating AWS resources. This should take around 5 minutes.
Error: Error creating cluster: stack-template.json size(=51673) exceeds the 51200 bytes limit of cloudformation. `--s3-uri s3://<bucket>/path/to/dir` must be specified to upload it to S3 beforehand
...

As the error message says, if you provide the `--s3-uri` option composed of a s3 bucket's name and a path to directory, `kube-aws` uploads your template to s3 if necessary:

```
$ kube-aws up --s3-uri s3://mybucket/mydir
```

resolves #38 (for now)