Reuse caching/encryption logic for the TLS bootstrap and auth tokens
This commit gets rid of the `token_config.go` file that were used to
read/cache/fingerprint the auth tokens file, and instead re-use the
same logic for reading/caching/fingerprinting TLS cert files.

This commit also changes the way the auth tokens file is assembled;
instead of storing the TLS bootstrap token in this file - and
extracting/encrypting it in every update (which cause the issue
reported in #633), we store the TLS bootstrap token separately in its
own file (which are cached and fingerprinted like the other ones) and
assemble the final auth tokens file in `cloud-config-controller`,
during the asset decryption step.