Drain interval implementation based on ASG lifecycle hooks
This commit configures a lifecycle hook to worker auto scaling groups
and a bash script that watches the current instance status: if the
instance is terminating, we run the node drainer script and let the
lifecycle hook time out, which causes the auto scaling group to
proceed to the next instance.

I could not find a way to set up a lifecycle hook without specifying a
SNS topic since apparently the `RoleARN` and `NotificationTargetARN`
attributes are required for the `AWS::AutoScaling::LifecycleHook`
resource.