Remove the old drainer system units
The idea is that voluntary node terminations should be preceded by a
`kubectl drain` executed by the cluster operator, and not leave this
to be done at the time the instance is terminated.

The main reason for this is that AWS do not wait for the drain command
to finish (especially if there are pod disruption budgets in place),
which renders this unit useless.