Generate certificates for control plane components
This commit is a first step towards disabling the apiserver insecure
port by changing kube-aws to generate two more client certs: one for
kube-controller-manager, and other for kube-scheduler.

Both certificates are generated with the correct common names for each
component, which makes RBAC to work out of the box; I added the
'--use-service-account-credentials' flag to controller manager so that
each controller can run with its own service account. [1]

Also, in order to allow those components to communicate securely with
the local apiserver instance over TLS, the apiserver certificate now
supports the '127.0.0.1' IP address as well.

[1] https://kubernetes.io/docs/admin/authorization/rbac/#controller-roles